# api/swagger/swagger.yaml
basePath: /api
info:
  contact:
    email: abdoelrachmad@gmail.com
    name: Maintainer
  description: Simple blog server built with go.
  title: GoBlog Server
  version: "1.0"
paths:
...
  /v1/auth/writer/comment/{uid}:
    delete:
      description: Delete a comment of my post (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete My Comment (Soft)
      tags:
      - Comment (Writer)
    get:
      description: Get a comment of mine.
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      content:
                        type: string
                      createdAt:
                        type: string
                      deletedAt:
                        type: string
                      email:
                        type: string
                      name:
                        type: string
                      parentCommentUid:
                        type: string
                      postUid:
                        type: string
                      replyCount:
                        type: integer
                      uid:
                        type: string
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get My Comment
      tags:
      - Comment (Writer)
  /v1/auth/writer/comment/{uid}/detrash:
    patch:
      description: Restore a deleted comment of my post (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore My Comment (Soft)
      tags:
      - Comment (Writer)
    put:
      description: Restore a deleted comment of my post (soft-deleted).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Restore My Comment (Soft)
      tags:
      - Comment (Writer)
  /v1/auth/writer/comment/{uid}/permanent:
    delete:
      description: Delete a comment of my post (permanent).
      parameters:
      - description: Comment's UID
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete My Comment (Permanent)
      tags:
      - Comment (Writer)
  /v1/auth/writer/comments:
    get:
      description: Get comments of mine.
      parameters:
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=active.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        content:
                          type: string
                        createdAt:
                          type: string
                        deletedAt:
                          type: string
                        email:
                          type: string
                        name:
                          type: string
                        parentCommentUid:
                          type: string
                        postUid:
                          type: string
                        replyCount:
                          type: integer
                        uid:
                          type: string
                      type: object
                  type: array
              type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get My Comments
      tags:
      - Comment (Writer)
  /v1/auth/writer/comments/stats:
    get:
      description: Get my comments's stats.
      parameters:
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=active.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  allOf:
                  - type: object
                  - properties:
                      currentPage:
                        type: integer
                      itemsPerPage:
                        type: integer
                      totalItems:
                        type: integer
                      totalPages:
                        type: integer
                    type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get My Comments Stats
      tags:
      - Comment (Writer)
...
  /v1/auth/writer/post/{uid}/comments:
    get:
      description: Get my post's comments.
      parameters:
      - description: Post's UID or slug
        in: path
        name: uid
        required: true
        type: string
      - description: Number of data to be shown.
        in: query
        name: show
        type: integer
      - description: Selected page of data.
        in: query
        name: page
        type: integer
      - description: Selected field to order data with.
        in: query
        name: order
        type: string
      - description: 'Ascending or descending, e.g.: ?asc=false.'
        in: query
        name: asc
        type: string
      - description: 'Filter data by type, e.g.: ?type=trash, ?type=active.'
        in: query
        name: type
        type: string
      produces:
      - application/json
      - application/msgpack
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - type: object
            - properties:
                data:
                  items:
                    allOf:
                    - type: object
                    - properties:
                        content:
                          type: string
                        createdAt:
                          type: string
                        deletedAt:
                          type: string
                        email:
                          type: string
                        name:
                          type: string
                        parentCommentUid:
                          type: string
                        postUid:
                          type: string
                        replyCount:
                          type: integer
                        uid:
                          type: string
                      type: object
                  type: array
              type: object
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - type: object
            - properties:
                message:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get My Post Comments
      tags:
      - Comment (Writer)
...
securityDefinitions:
  BearerAuth:
    description: '`Bearer <token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"